+++ e2e/tests/framework/pluginFixtures.e2e.spec.js
@@ -20,20 +20,25 @@
  * at runtime from the About dialog for additional information.
  *****************************************************************************/
 
-const { test, expect } = require('../pluginFixtures.js');
+/*
+This test suite is dedicated to testing our use of our custom fixtures to verify
+that they are working as expected.
+*/
+
+const { test, expect } = require('../../pluginFixtures.js');
 
 test.describe('pluginFixtures tests', () => {
     test.use({ domainObjectName: 'Timer' });
     let timerUUID;
 
-    test('Creates a timer object', ({ domainObject }) => {
+    test('Creates a timer object @framework', ({ domainObject }) => {
         const { uuid } = domainObject;
         const uuidRegexp = /[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/;
         expect(uuid).toMatch(uuidRegexp);
         timerUUID = uuid;
     });
 
-    test('Provides same uuid for subsequent uses of the same object', ({ domainObject }) => {
+    test('Provides same uuid for subsequent uses of the same object @framework', ({ domainObject }) => {
         const { uuid } = domainObject;
         expect(uuid).toEqual(timerUUID);
     });
